requirements.txt
asgiref==3.7.2
boto3==1.33.12
botocore==1.33.12
chardet==3.0.4
dj-database-url==0.5.0
Django==3.2.23
django-allauth==0.41.0
django-ckeditor==6.7.1
django-countries==7.2.1
django-crispy-forms==1.14.0
django-js-asset==2.2.0
django-storages==1.14.2
gunicorn==21.2.0
jmespath==1.0.1
oauthlib==3.2.2
Pillow==10.1.0
psycopg2==2.9.3
python3-openid==3.2.0
pytz==2023.3.post1
requests-oauthlib==1.3.1
s3transfer==0.8.2
sqlparse==0.4.4
stripe==8.1.0
urllib3==1.26.18
